backend/requirements.txt
fastapi>=0.120.0,<0.121.0
uvicorn>=0.38.0,<0.39.0
pymongo==4.6.0
motor==3.3.2
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt>=4.0.0
python-multipart>=0.0.9
python-dotenv>=1.0.1,<2.0.0
pydantic-settings>=2.5.2,<3.0.0
openai>=1.45.0,<2.0.0
starlette>=0.49.0,<0.50.0
anyio>=4.8.0,<5.0.0
tokenizers>=0.21,<0.22
langgraph>=0.2.0
langchain>=0.3.0
Pillow>=10.0.0
